fix: renumber states after spot purge
After calling purge_unreachable_states, it's necessary to renumber or remove states saved in external data structures, such as initial and final states, which are not managed by Spot's internal structures. To support this, Spot allows providing a custom state renumbering function when calling purge_unreachable_states. This function accepts a void* pointer, which we handle with a custom struct RenumberingContext containing pointers to the external initial and final states (which can be nullptr). We also define RenumberingContextVoidPtr as a void* alias to improve code readability, making it clear that a static cast to RenumberingContext* is needed.

This mechanism is used in two cases:

    To remove unreachable states from the initial NFA.
    To remove unreachable states after constructing the BackwardNFA.

Additionally, the handling of the dummy initial state has been refactored and fixed. The logic was moved into a new method, createDummyInitialStateWithEdgesToFinalStatesHavingPredecessors, which creates a dummy initial state pointing to final states that have at least one predecessor. After this, purge_unreachable_states is called to delete unreachable states, including final states without predecessors. The dummy initial state is no longer killed, so a counter for edges created from it is maintained. The totalStates() method now returns the total number of states minus 1 to exclude the dummy initial state. The totalEdges() method returns the total number of edges excluding those from the dummy initial state.