feat(abstraction): add TileExtractor class
The `TileExtractor` class is designed to extract all `Tile` objects from a given set of `Observable` instances. It provides methods to extract tiles from a vector of observables or a single observable. The class leverages the concepts of adjacency and border detection between polyhedra to efficiently identify and collect tiles from the given observables.